requirements.txt
astroid==3.3.11
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
fonttools==4.59.0
iniconfig==2.1.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.8.4
mccabe==0.7.0
networkx==3.2.1
numpy==2.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pygraphviz==1.11
pylint==3.3.7
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.17.0
tomlkit==0.13.3
